help.me

+++ src/Subject.java
@@ -2,18 +2,19 @@
 public class Subject {
 	
 	String name;
-	int grade;
+	double grade;
 
 	public Subject() {
 		
 	}
-	public Subject(String name, int grade) {
+	public Subject(String name, double grade) {
 		this.name = name;
 		this.grade = grade;
 	}
-	
+		
 	public void printInfo() {
 		System.out.println("name : " +this.name + "grade : " +this.grade);
 	}
+	
 
 }
